Move ActorIdFactory/ActorId out to separate header
Addresses long-standing TODO about moving the def out of io-channels.h
