+++ src/workerd/io/worker.h
@@ -15,6 +15,7 @@
 #include <workerd/jsg/async-context.h>
 #include <kj/mutex.h>
 #include <workerd/io/io-channels.h>
+#include <workerd/io/actor-id.h>
 #include <workerd/io/actor-storage.capnp.h>
 #include <workerd/io/request-tracker.h>
 #include <workerd/io/actor-cache.h>  // because we can't forward-declare ActorCache::SharedLru.
